// src/redux/tag-types.ts
/* eslint-disable no-unused-vars */
// eslint-disable-next-line no-shadow
export enum tagTypes {
  user = 'user',
  services = 'services',
  admin = 'admin',
  booking = 'booking',
  auth = 'auth',
  quiz = 'quiz',
  sets = 'sets',
  heardle = 'heardle',
  blog = 'blog',
  MetaData = 'MetaData',
}

export const tagTypesList = [
  tagTypes.user,
  tagTypes.auth,
  tagTypes.services,
  tagTypes.admin,
  tagTypes.booking,
  tagTypes.quiz,
  tagTypes.sets,
  tagTypes.heardle,
  tagTypes.blog,
  tagTypes.MetaData,
];